validate_service/requirements.txt
fastapi
uvicorn
python-multipart
boto3
Pillow
numpy
scikit-learn
albumentations
torch
torchvision
timm
onnx
onnxruntime
